// src/emailservice/src/schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/send_order_confirmation": {
    /** Sends an order confirmation email */
    post: {
      requestBody: {
        content: {
          "application/json": components["schemas"]["SendOrderConfirmationRequest"];
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: never;
        };
      };
    };
  };
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: {
    SendOrderConfirmationRequest: {
      email?: string;
      order?: components["schemas"]["OrderResult"];
    };
    OrderResult: {
      order_id?: string;
      shipping_tracking_id?: string;
      shipping_cost?: components["schemas"]["Money"];
      shipping_address?: components["schemas"]["Address"];
      items?: components["schemas"]["OrderItem"][];
    };
    OrderItem: {
      item?: components["schemas"]["CartItem"];
      cost?: components["schemas"]["Money"];
    };
    CartItem: {
      product_id?: string;
      quantity?: number;
    };
    Money: {
      currency_code?: string;
      units?: number;
      nanos?: number;
    };
    Address: {
      street_address?: string;
      city?: string;
      state?: string;
      country?: string;
      zip_code?: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export type operations = Record<string, never>;